refactor(dropdown-options): Extract common components.
