+++ lib/components/user/monitored-trip/trip-notifications-pane.tsx
@@ -1,10 +1,15 @@
 import { Alert, FormControl } from 'react-bootstrap'
 import { ExclamationTriangle } from '@styled-icons/fa-solid/ExclamationTriangle'
-import { Field, FormikProps } from 'formik'
-import { FormattedList, FormattedMessage, IntlShape, useIntl } from 'react-intl'
-import React, { Component, ComponentType, FormEvent, ReactNode } from 'react'
+import { FormattedList, FormattedMessage } from 'react-intl'
+import { FormikProps } from 'formik'
+import React, { Component, FormEvent } from 'react'
 import styled from 'styled-components'
 
+import {
+  DurationOptions,
+  Select,
+  YesNoOptions
+} from '../common/dropdown-options'
 import { FieldSet } from '../styled'
 import { IconWithText } from '../../util/styledIcon'
 import { MonitoredTrip } from '../types'
@@ -37,121 +42,6 @@ const Summary = styled.summary`
   font-weight: 700;
   margin-bottom: 5px;
 `
-
-/**
- * A label followed by a dropdown control.
- */
-const Select = ({
-  Control = FormControl,
-  children,
-  label,
-  name
-}: {
-  // Note the prop order required by typescript-sort-keys, also applied above.
-  Control?: ComponentType
-  children: ReactNode
-  label?: ReactNode
-  name: string
-}) => (
-  // <Field> is kept outside of <label> to accommodate layout in table/grid cells.
-  <>
-    {label && <label htmlFor={name}>{label}</label>}
-    <Field as={Control} componentClass="select" id={name} name={name}>
-      {children}
-    </Field>
-  </>
-)
-
-function Options({
-  defaultValue,
-  options
-}: {
-  defaultValue: number | string
-  options: { text: string; value: number | string }[]
-}) {
-  // <FormattedMessage> can't be used inside <option>.
-  const intl = useIntl()
-  return (
-    <>
-      {options.map(({ text, value }, i) => (
-        <option key={i} value={value}>
-          {value === defaultValue
-            ? intl.formatMessage(
-                { id: 'common.forms.defaultValue' },
-                { value: text }
-              )
-            : text}
-        </option>
-      ))}
-    </>
-  )
-}
-
-const basicYesNoOptions = [
-  {
-    id: 'yes',
-    value: 'true'
-  },
-  {
-    id: 'no',
-    value: 'false'
-  }
-]
-
-/**
- * Produces a yes/no list of options with the specified
- * default value (true for yes, false for no).
- */
-function YesNoOptions({ defaultValue }: { defaultValue: boolean }) {
-  // <FormattedMessage> can't be used inside <option>.
-  const intl = useIntl()
-  const options = basicYesNoOptions.map(({ id, value }) => ({
-    text:
-      id === 'yes'
-        ? intl.formatMessage({ id: 'common.forms.yes' })
-        : intl.formatMessage({ id: 'common.forms.no' }),
-    value
-  }))
-  return (
-    <Options
-      defaultValue={(defaultValue || false).toString()}
-      options={options}
-    />
-  )
-}
-
-/**
- * Produces a list of duration options with the specified default value.
- */
-function DurationOptions({
-  decoratorFunc,
-  defaultValue,
-  minuteOptions
-}: {
-  decoratorFunc?: (text: string, intl: IntlShape) => string
-  defaultValue: string | number
-  minuteOptions: number[]
-}) {
-  // <FormattedMessage> can't be used inside <option>.
-  const intl = useIntl()
-  const localizedMinutes = minuteOptions.map((minutes) => ({
-    text:
-      minutes === 60
-        ? intl.formatMessage({ id: 'components.TripNotificationsPane.oneHour' })
-        : intl.formatMessage(
-            { id: 'common.time.tripDurationFormat' },
-            { hours: 0, minutes, seconds: 0 }
-          ),
-    value: minutes
-  }))
-  const options = decoratorFunc
-    ? localizedMinutes.map(({ text, value }) => ({
-        text: decoratorFunc(text, intl),
-        value
-      }))
-    : localizedMinutes
-  return <Options defaultValue={defaultValue} options={options} />
-}
 
 interface Props extends FormikProps<MonitoredTrip> {
   notificationChannel: string
